1. 디버그 기능 추가 2. 경고 수정 3. 충돌 인터페이스 설정

+++ WinAPI/ContentsProject/BannerTextUI.cpp
@@ -53,7 +53,7 @@ void ABannerTextUI::SetTextOut(std::string_view _Text)
 		Renders[i]->SetActive(true);
 	}
 
-	for (int i = Text.size(); i < Renders.size(); i++)
+	for (size_t i = Text.size(); i < Renders.size(); i++)
 	{
 		Renders[i]->SetActive(false);
 	}
